FlagKeep

+++ src/github-apps.sh/109.sh
@@ -118,7 +118,9 @@ function installExecute
     # Wriete version
     AppsVersion "$app" "$FlagVersion"
 
-    echo rm "$file"
-    rm "$file"
+    if [[ $FlagKeep == 0 ]];then
+        echo rm "$file"
+        rm "$file"
+    fi
     echo "$success"
 }
